Refactor application
